Add tests, fix crate::rewrap function, free cluter in utils
* Add some more tests in src/tests.rs
* The space in `line.join` caused textwrap to wrap early in rewrap
  function
* Moved setup and cleanup functions to src/utils/term.rs
